Moving the solution to another folder
