+++ graphGenerator.js
@@ -1,21 +1,3 @@
-const Node =  class {
-  constructor(name, value) {
-    this.name = name;
-    this.value = value;
-    this.neighbors = [];
-  }
-
-  addNeighbors(neighbors){
-    if(neighbors){
-      this.neighbors = this.neighbors.concat(neighbors);
-    }
-
-    return this.neighbors;
-  }
-
-  getNeighbors(){
-    return this.neighbors;
-  }
+module.exports = class Node {
+  constructor() {}
 }
-
-module.exports = Node;
